+++ samples/pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
     sdk: flutter
   forui:
     path: ../forui
+  intl: ^0.19.0
   sugar: ^3.1.0
 
 dev_dependencies:
